requirements.txt
aiohttp~=3.11.6
aiogram~=3.17.0
python-dotenv